Use SDL2 to create windows and handle input. Fixes #48.
Launcher is now a .NET Core app, and creates its own SDL window.

Data viewer uses SDL_CreateWindowFrom to create an SDL window from an existing WinForms control, which in turn is embedded in the Eto.Forms-based viewer.

In both cases, SDL is used to handle keyboard and mouse input.