Fix color & team not being reset properly
Since they both use -1 as their dataItem value, the index needs to be shifted down by 1. The commit which set these to -1 in the first place seemed to be intentional, but if not a cleaner alternative would probably be to shift the indices so we don't have the -1 value anywhere.