Enable SaveFileTests to run locally
I don't think that these would be able to run in the CI builds at this stage as they would require game files, but this at least allows them to be run locally as a sanity check when making changes to sav parsing. Currently not all of the generals sav files past, but most do.